Adding conditional check for existence of composer.lock

+++ post-install.sh
@@ -1,9 +1,10 @@
 #!/bin/bash
+if [ ! -f "composer.lock" ]; then
+    cd `pwd`/vendor/jesgs/wordpress-dev/
 
-cd `pwd`/vendor/jesgs/wordpress-dev/
-
-cp index.php ../../../index.php
-cp .env.example ../../../.env.example
-cp php.ini ../../../php.ini
-cp phpinfo.php ../../../phpinfo.php
-cp wp-config.php ../../../wp-config.php
+    cp index.php ../../../index.php
+    cp .env.example ../../../.env.example
+    cp php.ini ../../../php.ini
+    cp phpinfo.php ../../../phpinfo.php
+    cp wp-config.php ../../../wp-config.php
+fi
